@@ -1,8 +1,8 @@
 /**
  * Angular 2 decorators and services
  */
-import { Component, OnInit, ViewEncapsulation } from "@angular/core";
-import { AppState } from "./app.service";
+import { Component, OnInit, ViewEncapsulation } from '@angular/core';
+import { AppState } from './app.service';
 
 /**
  * App Component
