@@ -1,4 +1,5 @@
-/**
- * App
- */
 export * from './app.module';
+export * from './home/home.module';
+export * from './listings/listings.module';
+export * from './listing-info/listing-info.module';
+export * from './side-nav/side-nav.module';
